+++ app/Http/Controllers/BookController.php
@@ -13,71 +13,81 @@ class BookController extends Controller
     protected $bookService;
     protected $authorService;
 
-    public function __construct(BookService $bookService, AuthorService $authorService){
+    public function __construct(BookService $bookService, AuthorService $authorService)
+    {
         $this->bookService = $bookService;
         $this->authorService = $authorService;
     }
 
-    public function index(){
-       
-        $books = $this->bookService->getAllBooks();
-        return view('books.index', compact('books'));
+    public function index()
+    {   
+        $books = $this->bookService->getAllBooks()->load('authors');
+
+        return response()->json($books);
     }
 
-    public function show($id){
+    public function show($id)
+    {
         $book = $this->bookService->getBookById($id);
-        return view('books.show', compact('book'));
+
+        if(!$book){
+            return response()->json(['message' => 'Book not found'], 404);
+        }
+
+        $book = $book->load('authors');
+
+        return response()->json($book);
     }
 
 
-    public function create(){
-    
-        $authors = $this->authorService->getAllAuthors(); 
-        return view('books.create', compact('authors'));
-    }
-
-    public function store(Request $request){
-  
-
+    public function store(Request $request)
+    {
         $this->validate($request, [
             'title' => 'required|string|max:255|unique:books',
             'authors' => 'required|array'
         ]);
     
-        $book = $this->bookService->createBook($request->only(['title']));
-        $book->authors()->attach($request->authors); 
+        $books = $this->bookService->createBook($request->only(['title']));
 
-        return redirect()->route('books.index');
+
+
+        $books->authors()->attach($request->authors); // take the authors and attach them to the book by ading them to the pivot table
+
+        return response()->json(['message' => 'Book created successfully', 'book' => $books], 202);
     }
 
-    public function edit($id){
+
+    public function update($id, Request $request)
+    {
         $book = $this->bookService->getBookById($id);
-        $authors = $this->authorService->getAllAuthors();
-        $selectedAuthors = $book->authors->pluck('id')->toArray();
-        return view('books.edit', compact('book', 'authors', 'selectedAuthors'));
- 
-    }
 
-    public function update($id, Request $request){
+        if(!$book){
+            return response()->json(['message' => 'Book not found'], 404);
+        }
+
         $this->validate($request, [
             'title' => 'required|string|max:255|unique:books,title,' . $id,
             'authors' => 'required|array'
         ]);
 
         $this->bookService->updateBook($id, $request->only(['title']));
-        $book = $this->bookService->getBookById($id);
-        $book->authors()->sync($request->authors);
-        return redirect()->route('books.index');
 
+      
+        $book->authors()->sync($request->authors); // can also be a function in the bookService
+        
+        return response()->json(['message' => 'Book updated successfully', 'data' => $book], 200);
     }
 
-    public function destroy($id){
-        DB::table('author_book')->where('book_id', $id)->delete();
+    public function destroy($id)
+    {
+        $book = $this->bookService->getBookById($id);
 
-        DB::table('user_book')->where('book_id', $id)->delete();
+        if (!$book) {
+            return response()->json(['message' => 'Book not found'], 404);
+        }
 
         $this->bookService->deleteBook($id);
 
-        return redirect()->route('books.index');
+        return response()->json(['message' => 'Book deleted successfully'], 200);
     }
 }
